Rename agent_telemetry -> client_telemetry
