+++ client_telemetry/pcap_stats_test.go
@@ -1,4 +1,4 @@
-package agent_telemetry
+package client_telemetry
 
 import (
 	"encoding/json"
